isCustomDate state removed
