+++ packages/cli/src/editor/components/RenderModal/RenderModalData.tsx
@@ -22,6 +22,11 @@ import {SegmentedControl} from '../SegmentedControl';
 import type {TypeCanSaveState} from './get-render-modal-warnings';
 import {getRenderModalWarnings} from './get-render-modal-warnings';
 import {RenderModalJSONPropsEditor} from './RenderModalJSONPropsEditor';
+import type {SerializedJSONWithDate} from './SchemaEditor/date-serialization';
+import {
+	deserializeJSONWithDate,
+	serializeJSONWithDate,
+} from './SchemaEditor/date-serialization';
 import {SchemaEditor} from './SchemaEditor/SchemaEditor';
 import {
 	NoDefaultProps,
@@ -30,6 +35,18 @@ import {
 import {WarningIndicatorButton} from './WarningIndicatorButton';
 
 type Mode = 'json' | 'schema';
+
+export type State =
+	| {
+			str: string;
+			value: unknown;
+			validJSON: true;
+	  }
+	| {
+			str: string;
+			validJSON: false;
+			error: string;
+	  };
 
 export type PropsEditType = 'input-props' | 'default-props';
 
@@ -73,6 +90,15 @@ const tabWrapper: React.CSSProperties = {
 
 const persistanceKey = 'remotion.show-render-modalwarning';
 
+const parseJSON = (str: string): State => {
+	try {
+		const value = deserializeJSONWithDate(str);
+		return {str, value, validJSON: true};
+	} catch (e) {
+		return {str, validJSON: false, error: (e as Error).message};
+	}
+};
+
 const getPersistedShowWarningState = () => {
 	const val = localStorage.getItem(persistanceKey);
 	if (!val) {
@@ -103,7 +129,6 @@ export const RenderModalData: React.FC<{
 }) => {
 	const [mode, setMode] = useState<Mode>('schema');
 	const [valBeforeSafe, setValBeforeSafe] = useState<unknown>(inputProps);
-	const [isCustomDateUsed, setIsCustomDateUsed] = useState<boolean>();
 	const [saving, setSaving] = useState(false);
 	const zodValidationResult = useMemo(() => {
 		return composition.schema.safeParse(inputProps);
@@ -111,6 +136,17 @@ export const RenderModalData: React.FC<{
 	const [showWarning, setShowWarningWithoutPersistance] = useState<boolean>(
 		() => getPersistedShowWarningState()
 	);
+
+	const serializedJSON: SerializedJSONWithDate = useMemo(() => {
+		const value = inputProps ?? {};
+		return serializeJSONWithDate(value, 2);
+	}, [inputProps]);
+
+	const isCustomDateUsed = serializedJSON.customDateUsed;
+	const [localValue, setLocalValue] = React.useState<State>(() => {
+		return parseJSON(serializedJSON.serializedString);
+	});
+
 	const cliProps = getInputProps();
 	const [canSaveDefaultProps, setCanSaveDefaultProps] =
 		useState<TypeCanSaveState>({
@@ -293,7 +329,9 @@ export const RenderModalData: React.FC<{
 					onSave={onUpdate}
 					valBeforeSafe={valBeforeSafe}
 					showSaveButton={showSaveButton}
-					setIsCustomDateUsed={setIsCustomDateUsed}
+					localValue={localValue}
+					setLocalValue={setLocalValue}
+					parseJSON={parseJSON}
 				/>
 			)}
 		</div>
